requirements.txt
absl-py==0.7.1
alabaster==0.7.12
astor==0.8.0
attrs==19.1.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29.10
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.7
graphviz==0.13.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
imagesize==1.2.0
imgaug==0.2.9
implicit==0.4.0
imutils==0.5.3
ipykernel==5.1.2
ipyparallel==6.2.4
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.1
jsonpickle==1.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
labelme==3.16.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nose==1.3.7
notebook==6.0.1
numpy==1.16.4
object-detection==0.1
opencv-python==4.1.0.25
opencv-python-headless==4.1.2.30
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
pyaml==19.12.0
pycocotools==2.0
pydensecrf==1.0rc2
pydot==1.4.1
pygame==1.9.6
Pygments==2.4.2
pyparsing==2.4.0
PyQt5==5.13.0
PyQt5-sip==12.7.0
pyqt5-tools==5.13.0.1.5
pyreadline==2.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.1.0
qtconsole==4.5.5
QtPy==1.7.1
requests==2.22.0
schema==0.7.1
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.3.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spline==1.12
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
